Refactor: Split utility.py into focused utils/* modules (#1608)
Move embed types/classes to utils/embeds.py (EmbedColor, StatusIcon,
  ErrorEmbedCategory, TanjunEmbed, embed builders)
- Move math/level helpers to utils/math.py (NumericStringParser,
  eval_expr, level/XP calculations)
- Move time helpers to utils/time.py (relative time string parsing)
- Move HTTP helpers to utils/http.py (getGif, ImgBB/bytebin upload)
- Move GitHub helpers to utils/github.py (missingLocalization,
  addFeedback)
- Keep utility.py as a backward-compatible re-export module with
  SafeInteraction, DiscordSafe, and misc helpers
- Fix disable_level_system.py to use categorized embed builders
  explicitly
